% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESMAtransRateReader.R
\name{ESMAtransRateReader}
\alias{ESMAtransRateReader}
\title{Transition rate matrix reader from ESMA website}
\usage{
ESMAtransRateReader(filepath)
}
\arguments{
\item{filepath}{The character string of the path at which the .csv file is stored.}
}
\value{
list
\describe{
  \item{title}{The name of the CRA}
  \item{mat}{The transition rate matrix}
}
}
\description{
Reads a .csv file in the given format available at \url{https://cerep.esma.europa.eu/cerep-web/statistics/transitionMatrice.xhtml}
("CSV export") and parses it. The content of such file includes the name of the credit rating agency, the period and time horizon
of statistics, the cardinal matrix of transitions and the transition rate matrix.
}
\examples{
\dontrun{
  ESMAtransRateReader("cerep878219543695070733.csv")
  ESMAtransRateReader(system.file(
                       'extdata', "cerepJCRA.csv"),
                       package = 'mctools')
}

}
