% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{GUI}
\alias{GUI}
\title{Shiny App}
\usage{
GUI()
}
\description{
Wraps the start function of shiny app for Markov Chain analysis. Check
\url{https://github.com/unimi-dse/5547f4ce/blob/master/README.md}
for a complete description of available tools.
}
