% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_gen.R
\name{mat_gen}
\alias{mat_gen}
\title{Transition matrix generator}
\usage{
mat_gen(n, p)
}
\arguments{
\item{n}{The dimension (nrow = ncol) of the matrix to be created}

\item{p}{The probability that one generic element of the matrix will be set to 0.}
}
\description{
Generate a transition matrix (square matrix whose rows sum 1), given its
dimension and a probability of non-connection.
}
\note{
Entering a probability of non-connection overcomes the problem that, when
generating a continuosly and uniformly distributed sequence, the probability to
obtain 0 is 0.
}
\examples{
mat_gen(4,0.2)

}
