% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_val.R
\name{mat_val}
\alias{mat_val}
\title{Transition matrix validation}
\usage{
mat_val(M)
}
\arguments{
\item{M}{A candidate transition matrix.}
}
\value{
A character string specifing the cause of invalidation, or the matrix itself if it is valid.
}
\description{
Validates a transition matrix. Checks that each entry of the matrix is a non-negative
number and that each row add up to 1.
}
\examples{
A <- matrix(c(0.33, 0.66, 0.00, 0.00,
               0.16, 0.17, 0.66, 0.00,
               0.50, 0.00, 0.50, 0.00,
               0.00, 0.00, 0.00, 0.00), nrow = 4, byrow = TRUE)
mat_val(A)
mat_val(matrixBuilder(A))

}
