% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixBuilder.R
\name{matrixBuilder}
\alias{matrixBuilder}
\title{Transition martix builder from inexact transition rate matrix}
\usage{
matrixBuilder(T)
}
\arguments{
\item{T}{A transition rate matrix}
}
\value{
The adjusted transition matrix
}
\description{
Adjusts an inexact transition matrix, which may have zero-rows and zero-columns and/or may
showcase sums of rows that slightly differ from 1. This two aspects may
occur as a result of parsing a transition rate matrix as given by ESMA website (see
\code{\link{ESMAtransRateReader}}), where certain credit ratings are never experienced
and decimal representation of fractions leads to approximation errors.
}
\examples{
 A <- matrix(c(0.33, 0.66, 0.00, 0.00,
               0.16, 0.17, 0.66, 0.00,
               0.50, 0.00, 0.50, 0.00,
               0.00, 0.00, 0.00, 0.00), nrow = 4, byrow = TRUE)
 matrixBuilder(A)

}
