% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percToNum.R
\name{percToNum}
\alias{percToNum}
\title{Percentage to numeric}
\usage{
percToNum(x)
}
\value{
The numeric vector or matrix resulting. \code{dimnames} is left unchanged.
}
\description{
Turns a character vector or matrix of percentages into numeric (dividing by 100).
}
