% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRows.R
\name{printRows}
\alias{printRows}
\title{Matrix rows print}
\usage{
printRows(M)
}
\arguments{
\item{M}{A matrix}
}
\value{
A string containing the rows of M.
}
\description{
Chains the rows of a matrix into a single string in the format
"\code{V1 = \{...\}, V2 = \{...\}, ...}".
}
\examples{
printRows(round(mat_gen(3,0.2),3))

}
