% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printVecNames.R
\name{printVecNames}
\alias{printVecNames}
\title{Vector print}
\usage{
printVecNames(v)
}
\arguments{
\item{v}{A vector to be printed}
}
\value{
A string containing the elements of the vector separated by "\code{,}"
}
\description{
Chains the elements of a vector along with their names in a string in the format
\code{"name1 = val1, name2 = val2, ..."}.
}
\examples{
printVecNames(c("State 1" = "sunny", "State 2" = "cloudy", "State 3" = "rainy"))

}
